requirements.txt
aniso8601==9.0.1
argon2-cffi==20.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
cassandra-driver==3.25.0
catboost==0.26.1
category-encoders==2.2.2
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.1
Flask-RESTful==0.3.9
geomet==0.2.1.post1
graphviz==0.17
gunicorn==20.1.0
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.2.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywinpty==1.1.3
pyzmq==22.2.1
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
statsmodels==0.12.2
sympy==1.8
tenacity==8.0.1
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
xgboost==1.4.2
